Every page fully generated with js
Every page is fully interactive and responsive and also changed how subscribers are saved and now they are loaded in the sidebar. The things left are adding account management in the header and whenever there is video when clicked to sent to the video page. Maybe add watch later page.