Made dislike button and new things
Now there is dislike button in the video page that does not show its counter because its a youtube copy. In the liked videos page now there is a option to remove a video from the users liked videos. Now this option will be added to watch later and for last make the dots interactive in the video page for the side videos and in the You page.